Fix display issue with the banner
Because of that the symlink still tries to load the banner.jpg in the same directory, the banner have not become displayed. This commit fixes this issue by creating a symlink to the banner at the root of the repository. Means now the README for the .github repository and for public profile is automaticly synced with the banner and README at the root directory.